Change "Import" tooltip on pool overview page
The "Import" tooltip  on the pool overview page can be understood as referring to `zfs import`, especially since it is right next to an "add/create" button.

See for instance the discussion in [0] which matches my own experience: After starting up the zfs-plugin, I tried pushing "import" to import my existing zfs pool. When nothing happened, I assumed this was due to things still being beta on OMV6 and went to `zfs import` the pool manually from terminal. This made the pool show up correctly on the overview page, but the filesystems were not listed. 
Assuming that the "import" button was only related to `zfs import` I never thought to press it again. A contributing cause was that the zfs-page looked fine, and I assumed to missing filesystem listings to be a problem outside the zfs-plugin.

In the edit above, I have suggested "Update OMV filesystem list" instead of "Import", but I don't know the exact semantics of the action so there is probably a better phrase -- feel free to use any better wording you can think of.

[0]: https://forum.openmediavault.org/index.php?thread%2F41263-zfs-on-omv6%2F=&postID=306779#post306779=